+++ codegen/src/main/java/io/improbable/keanu/codegen/python/TemplateProcessor.java
@@ -3,6 +3,7 @@ package io.improbable.keanu.codegen.python;
 import freemarker.template.Configuration;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
+import lombok.experimental.UtilityClass;
 
 import java.io.File;
 import java.io.FileWriter;
@@ -13,9 +14,10 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.Map;
 
-class TemplateProcessor {
+@UtilityClass
+public class TemplateProcessor {
 
-    static void processDataModel(Map<String, Object> dataModel, Template fileTemplate, Writer fileWriter) {
+    public void processDataModel(Map<String, Object> dataModel, Template fileTemplate, Writer fileWriter) {
         try {
             fileTemplate.process(dataModel, fileWriter);
             fileWriter.close();
@@ -26,7 +28,7 @@ class TemplateProcessor {
         }
     }
 
-    static Template getFileTemplate(String templateFile) {
+    public Template getFileTemplate(String templateFile) {
         Configuration cfg = new Configuration(Configuration.VERSION_2_3_28);
         cfg.setClassForTemplateLoading(Runner.class, "/");
         try {
@@ -36,7 +38,7 @@ class TemplateProcessor {
         }
     }
 
-    static Writer createFileWriter(String fileToWrite) {
+    public Writer createFileWriter(String fileToWrite) {
         try {
             Files.deleteIfExists(Paths.get(fileToWrite));
             return new FileWriter(new File(fileToWrite), true);